Migrate graph data objects to gen-core
