+++ gen-graph/src/lib.rs
@@ -1,6 +1,6 @@
 use std::{
     collections::{HashMap, HashSet, VecDeque},
-    fmt::{self, Debug},
+    fmt::Debug,
     hash::Hash,
     iter::from_fn,
     rc::Rc,
@@ -24,57 +24,10 @@ use petgraph::{
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 
-pub mod traits;
-pub use traits::MergeGraph;
 pub mod graph_loader;
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash, Ord, PartialOrd, Deserialize, Serialize)]
-pub struct GraphNode {
-    pub node_id: HashId,
-    pub sequence_start: i64,
-    pub sequence_end: i64,
-}
-
-impl fmt::Display for GraphNode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            f,
-            "{}[{}-{}]",
-            self.node_id, self.sequence_start, self.sequence_end
-        )
-    }
-}
-
-impl GraphNode {
-    pub fn length(&self) -> i64 {
-        self.sequence_end - self.sequence_start
-    }
-}
-
-/// A slice of a single graph block: the block itself plus local start/end byte
-/// offsets within that block's sequence.  Middle blocks in a multi-block locus
-/// span the full block (`start = 0`, `end = block.length()`).
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct GraphNodeSlice {
-    pub block: GraphNode,
-    /// Local start offset within the block's sequence slice (`0..block.length()`).
-    pub start: usize,
-    /// Local end offset, exclusive (`start..=block.length()`).
-    pub end: usize,
-    /// 5'→3' orientation of this slice.
-    pub strand: Strand,
-}
-
-impl GraphNodeSlice {
-    pub fn full(block: GraphNode, strand: Strand) -> Self {
-        Self {
-            block,
-            start: 0,
-            end: block.length() as usize,
-            strand,
-        }
-    }
-}
+pub mod traits;
+pub use gen_core::{GraphNode, GraphNodePosition, GraphNodeSlice};
+pub use traits::MergeGraph;
 
 pub type GenGraph = DiGraphMap<GraphNode, Vec<GraphEdge>>;
 pub type OperationGraph = DiGraphMap<HashId, ()>;
@@ -87,20 +40,6 @@ pub struct GraphEdge {
     pub chromosome_index: i64,
     pub phased: i64,
     pub created_on: i64,
-}
-
-/// A cursor within a GraphNode
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash, Ord, PartialOrd, Deserialize, Serialize)]
-pub struct GraphNodePosition {
-    pub graph_node: GraphNode,
-    /// Distance from sequence_start of graph_node
-    pub offset: i64,
-}
-
-impl GraphNodePosition {
-    pub fn coordinate(&self) -> i64 {
-        self.graph_node.sequence_start + self.offset
-    }
 }
 
 #[derive(Debug, Error)]
